+++ frontend/views/user/create.php
@@ -19,7 +19,6 @@ $this->params['breadcrumbs'][] = $this->title;
 
     <?= $this->render('_form', [
         'model' => $model,
-        'isCurrentUser' => false,
         'showDeleteButton' => false
     ]) ?>