+++ frontend/views/user/view.php
@@ -6,7 +6,6 @@
     use yii\widgets\DetailView;
 
     /** @var yii\web\View $this */
-    /** @var bool $isCurrentUser */
     /** @var common\models\User $model */
 
     $this->title = $model->username;
@@ -27,14 +26,14 @@
     <p>
         <?= Html::a(Yii::t('views', 'Edit'), ['update', 'id' => $model->id], ['class' => 'btn btn-success']) ?>
 
-        <?php if ($isCurrentUser): ?>
+        <?php if ($model->getIsCurrentUser()): ?>
 
             <span class="d-inline-block" tabindex="0" title="<?= Yii::t('controllers', 'You cannot delete currently logged-in user.') ?>" data-bs-placement="right">
 
         <?php endif; ?>
 
         <?= Html::a(Yii::t('views', 'Delete'), ['delete', 'id' => $model->id], [
-            'class' => 'btn btn-danger delete-user-btn' . ($isCurrentUser ? ' disabled' : ''),
+            'class' => 'btn btn-danger delete-user-btn' . ($model->getIsCurrentUser() ? ' disabled' : ''),
             'data' => [
                 'modal-username' => Html::encode($model->username),
                 'modal-type' => Html::encode($model->getAttributeDesc('type')),
@@ -42,7 +41,7 @@
             ],
         ]) ?>
 
-        <?php if ($isCurrentUser): ?></span><?php endif; ?>
+        <?php if ($model->getIsCurrentUser()): ?></span><?php endif; ?>
 
     </p>
 
@@ -114,7 +113,7 @@
 
     <?php endif; ?>
 
-    <?php if ($isCurrentUser): ?>
+    <?php if ($model->getIsCurrentUser()): ?>
 
         <br /><?= Yii::t('controllers', 'You cannot delete currently logged-in user.') ?>