+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ### Bugs
 
-- [#33](https://github.com/niner-games/magiedit.web/issues/33): Registration form fails to complete ([Mr. Fat Guy](https://github.com/mrfatguy))
+- [#34](https://github.com/niner-games/magiedit.web/issues/34): Downgrading yourself to non-admin throws 403 exception ([Mr. Fat Guy](https://github.com/mrfatguy))
 - [#28](https://github.com/niner-games/magiedit.web/issues/28): Row's click event overrides buttons' events ([Mr. Fat Guy](https://github.com/mrfatguy))
 - [#27](https://github.com/niner-games/magiedit.web/issues/27): Wrong field for user type validation error ([Mr. Fat Guy](https://github.com/mrfatguy))
+- [#33](https://github.com/niner-games/magiedit.web/issues/33): Registration form fails to complete ([Mr. Fat Guy](https://github.com/mrfatguy))
 
 ### Features